+++ src/mem.h
@@ -3,6 +3,6 @@
 #ifndef MEM_H
 #define MEM_H
 
-void memcpy(void* dest, const void* src, size_t n);
+void memcpy(void* dest, const void* src, long unsigned int n);
 
 #endif
